New: Show detailed queue status on Calendar and Series Details
Closes #3775
