+++ frontend/src/Activity/Queue/QueueDetails.css
@@ -1,0 +1,5 @@
+.progressBarContainer {
+  display: flex;
+  justify-content: center;
+  width: 100%;
+}
