+++ frontend/src/Activity/Queue/QueueStatus.css
@@ -1,0 +1,3 @@
+.noMessages {
+  margin-bottom: 10px;
+}
