+++ frontend/src/Calendar/Events/CalendarEventQueueDetails.js
@@ -12,6 +12,7 @@ function CalendarEventQueueDetails(props) {
     status,
     trackedDownloadState,
     trackedDownloadStatus,
+    statusMessages,
     errorMessage
   } = props;
 
@@ -26,16 +27,15 @@ function CalendarEventQueueDetails(props) {
       status={status}
       trackedDownloadState={trackedDownloadState}
       trackedDownloadStatus={trackedDownloadStatus}
+      statusMessages={statusMessages}
       errorMessage={errorMessage}
       progressBar={
-        <div title={`Episode is downloading - ${progress.toFixed(1)}% ${title}`}>
-          <CircularProgressBar
-            progress={progress}
-            size={20}
-            strokeWidth={2}
-            strokeColor={'#7a43b6'}
-          />
-        </div>
+        <CircularProgressBar
+          progress={progress}
+          size={20}
+          strokeWidth={2}
+          strokeColor={'#7a43b6'}
+        />
       }
     />
   );
@@ -49,6 +49,7 @@ CalendarEventQueueDetails.propTypes = {
   status: PropTypes.string.isRequired,
   trackedDownloadState: PropTypes.string.isRequired,
   trackedDownloadStatus: PropTypes.string.isRequired,
+  statusMessages: PropTypes.arrayOf(PropTypes.object),
   errorMessage: PropTypes.string
 };
